Many small patches
- Restructure Config page so UI load happens in one step
- Move to one ButtonLink element for sidebar additions
- Add pointer to main dashboard to all servers
- Add Maps to director, both server level and overview
  - Overview map is very introductory
