+++ web_ui/frontend/app/cache/layout.tsx
@@ -18,13 +18,15 @@
 
 import { Box, Tooltip } from '@mui/material';
 
-import { Sidebar } from '@/components/layout/Sidebar';
+import { ButtonLink, Sidebar } from '@/components/layout/Sidebar';
 import Link from 'next/link';
 import Image from 'next/image';
 import PelicanLogo from '@/public/static/images/PelicanPlatformLogo_Icon.png';
 import IconButton from '@mui/material/IconButton';
 import BuildIcon from '@mui/icons-material/Build';
 import Main from '@/components/layout/Main';
+import { PaddedContent } from '@/components/layout';
+import { Dashboard, MapOutlined } from '@mui/icons-material';
 
 export const metadata = {
   title: 'Pelican Cache',
@@ -39,17 +41,16 @@ export default function RootLayout({
   return (
     <Box display={'flex'} flexDirection={'row'}>
       <Sidebar>
-        <Box pt={1}>
-          <Tooltip title={'Config'} placement={'right'}>
-            <Link href={'/config/'}>
-              <IconButton>
-                <BuildIcon />
-              </IconButton>
-            </Link>
-          </Tooltip>
-        </Box>
+        <ButtonLink title={'Cache'} href={'/cache'}>
+          <Dashboard />
+        </ButtonLink>
+        <ButtonLink title={'Config'} href={'/config/'}>
+          <BuildIcon />
+        </ButtonLink>
       </Sidebar>
-      <Main>{children}</Main>
+      <Main>
+        <PaddedContent>{children}</PaddedContent>
+      </Main>
     </Box>
   );
 }
