+++ web_ui/frontend/app/config/components/index.ts
@@ -1,0 +1,2 @@
+export * from './ConfigDisplay';
+export * from './TableOfContents';
