+++ web_ui/frontend/app/director/layout.tsx
@@ -18,13 +18,14 @@
 
 import { Box, Tooltip } from '@mui/material';
 
-import { Sidebar } from '@/components/layout/Sidebar';
+import { ButtonLink, Sidebar } from '@/components/layout/Sidebar';
 import Link from 'next/link';
 import Image from 'next/image';
 import PelicanLogo from '@/public/static/images/PelicanPlatformLogo_Icon.png';
 import IconButton from '@mui/material/IconButton';
 import BuildIcon from '@mui/icons-material/Build';
 import Main from '@/components/layout/Main';
+import { Dashboard, MapOutlined } from '@mui/icons-material';
 
 export const metadata = {
   title: 'Pelican Director',
@@ -39,15 +40,15 @@ export default function RootLayout({
   return (
     <Box display={'flex'} flexDirection={'row'}>
       <Sidebar>
-        <Box pt={1}>
-          <Tooltip title={'Config'} placement={'right'}>
-            <Link href={'/config/'}>
-              <IconButton>
-                <BuildIcon />
-              </IconButton>
-            </Link>
-          </Tooltip>
-        </Box>
+        <ButtonLink title={'Dashboard'} href={'/director'}>
+          <Dashboard />
+        </ButtonLink>
+        <ButtonLink title={'Map'} href={'/director/map/'}>
+          <MapOutlined />
+        </ButtonLink>
+        <ButtonLink title={'Config'} href={'/config/'}>
+          <BuildIcon />
+        </ButtonLink>
       </Sidebar>
       <Main>{children}</Main>
     </Box>
