+++ web_ui/frontend/app/page.module.css
@@ -120,11 +120,11 @@
   position: absolute;
   top: 0;
   left: 0;
-  width: 120%;
+  width: 75px;
   height: 120%;
   margin-left: -5%;
   margin-top: -1%;
-  background-color: white;
+  background-color: rgba(255, 255, 255, 0.95);
   transition: opacity 0.5s ease-out;
   opacity: 98%;
 }
