+++ web_ui/frontend/app/registry/components/CustomRegistrationField/index.tsx
@@ -1,4 +1,4 @@
-import type { CustomRegistrationField } from '@/components/Config/index.d';
+import type { CustomRegistrationField } from '@/components/Config/index';
 import { BooleanField } from './BooleanField';
 import { ErrorField } from './ErrorField';
 import { StringField } from './StringField';
