+++ web_ui/frontend/components/Config/MultiSelectField.tsx
@@ -13,7 +13,7 @@ import React, {
   ChangeEvent,
 } from 'react';
 
-import { ParameterInputProps } from '@/components/Config/index.d';
+import { ParameterInputProps } from '@/components/Config/index';
 import { createId, buildPatch } from './util';
 import OutlinedInput from '@mui/material/OutlinedInput';
