+++ web_ui/frontend/components/Config/ObjectField/InstitutionForm.tsx
@@ -1,4 +1,4 @@
-import { Institution } from '@/components/Config/index.d';
+import { Institution } from '@/components/Config/index';
 import React from 'react';
 import { Box, Button, TextField } from '@mui/material';
