+++ web_ui/frontend/components/Config/ObjectField/LotForm.tsx
@@ -1,4 +1,4 @@
-import { Action, Lot, Path } from '@/components/Config/index.d';
+import { Action, Lot, Path } from '@/components/Config/index';
 import React from 'react';
 import { Box, Button } from '@mui/material';
