+++ web_ui/frontend/components/Config/ObjectField/OIDCAuthenticationRequirementForm.tsx
@@ -1,4 +1,4 @@
-import { OIDCAuthenticationRequirement } from '@/components/Config/index.d';
+import { OIDCAuthenticationRequirement } from '@/components/Config/index';
 import React from 'react';
 import { Box, Button, TextField } from '@mui/material';
