+++ web_ui/frontend/components/Config/ObjectField/PathForm.tsx
@@ -1,4 +1,4 @@
-import { Path } from '@/components/Config/index.d';
+import { Path } from '@/components/Config/index';
 import React from 'react';
 import { Box, Button } from '@mui/material';
