+++ web_ui/frontend/components/Config/StringField.tsx
@@ -12,7 +12,7 @@ import React, {
   ChangeEvent,
 } from 'react';
 
-import { ParameterInputProps } from '@/components/Config/index.d';
+import { ParameterInputProps } from '@/components/Config/index';
 import { createId, buildPatch } from './util';
 
 export type StringFieldProps = {
