+++ web_ui/frontend/components/Config/VerifiableTextField.tsx
@@ -12,7 +12,7 @@ import React, {
   ChangeEvent,
 } from 'react';
 
-import { ParameterInputProps, Duration } from '@/components/Config/index.d';
+import { ParameterInputProps, Duration } from '@/components/Config/index';
 import { createId, buildPatch } from './util';
 
 export type VerifiableParameter = ParameterInputProps & {
