+++ web_ui/frontend/components/Map/index.ts
@@ -1,1 +1,3 @@
-export * from "./SinglePointMap"
+export * from './SinglePointMap';
+export * from './DefaultMap';
+export * from './ServerMap';
