+++ web_ui/frontend/components/layout/PaddedContent.tsx
@@ -1,0 +1,10 @@
+import { Box } from '@mui/material';
+import { ReactNode } from 'react';
+
+export const PaddedContent = ({ children }: { children: ReactNode }) => {
+  return (
+    <Box p={2} flexGrow={1}>
+      {children}
+    </Box>
+  );
+};
