+++ web_ui/frontend/components/layout/index.ts
@@ -1,0 +1,7 @@
+export * from './AuthenticatedContent';
+export * from './Drawer';
+export * from './Header';
+export * from './Main';
+export * from './OriginSidebar';
+export * from './SidebarSpeedDial';
+export * from './PaddedContent';
